requirements.txt
python-telegram-bot
aiohttp
beautifulsoup4
flask   
quart 